+++ database/migrations/2023_05_23_151550_create_document_items_table.php
@@ -19,11 +19,11 @@ return new class extends Migration
             $table->string('type');
             $table->string('name');
             $table->text('description')->nullable();
-            $table->decimal('quantity', 7, 2);
-            $table->decimal('price', 15, 4);
+            $table->double('quantity', 7, 2);
+            $table->double('price', 15, 4);
             $table->foreignId('tax_id')->nullable()->constrained()->nullOnDelete();
             $table->foreignId('discount_id')->nullable()->constrained()->nullOnDelete();
-            $table->decimal('total', 15, 4);
+            $table->double('total', 15, 4);
             $table->foreignId('created_by')->nullable()->constrained('users')->nullOnDelete();
             $table->foreignId('updated_by')->nullable()->constrained('users')->nullOnDelete();
             $table->timestamps();
